cleanup legacy client
