+++ packages/prelude/package.json
@@ -151,16 +151,6 @@
         "default": "./_cjs/client/clientFor.cjs"
       }
     },
-    "./client/clientFor2": {
-      "import": {
-        "types": "./dist/client/clientFor2.d.ts",
-        "default": "./dist/client/clientFor2.js"
-      },
-      "require": {
-        "types": "./dist/client/clientFor2.d.ts",
-        "default": "./_cjs/client/clientFor2.cjs"
-      }
-    },
     "./client/config": {
       "import": {
         "types": "./dist/client/config.d.ts",
@@ -179,26 +169,6 @@
       "require": {
         "types": "./dist/client/errors.d.ts",
         "default": "./_cjs/client/errors.cjs"
-      }
-    },
-    "./client/fetch": {
-      "import": {
-        "types": "./dist/client/fetch.d.ts",
-        "default": "./dist/client/fetch.js"
-      },
-      "require": {
-        "types": "./dist/client/fetch.d.ts",
-        "default": "./_cjs/client/fetch.cjs"
-      }
-    },
-    "./client/router": {
-      "import": {
-        "types": "./dist/client/router.d.ts",
-        "default": "./dist/client/router.js"
-      },
-      "require": {
-        "types": "./dist/client/router.d.ts",
-        "default": "./_cjs/client/router.cjs"
       }
     },
     "./faker": {
