+++ packages/prelude/src/client.ts
@@ -2,5 +2,4 @@
 export * from "./client/clientFor.js"
 export * from "./client/config.js"
 export * from "./client/errors.js"
-export * from "./client/fetch.js"
 // codegen:end
