+++ packages/schema/src/index.ts
@@ -2,8 +2,6 @@ export * from "@effect/schema/Schema"
 
 export * from "./Class.js"
 export { Class, TaggedClass } from "./Class.js"
-export * as REST from "./REST.js"
-export { Delete, Get, Patch, Post, Put, Req } from "./REST.js"
 
 export { fromBrand, nominal } from "./brand.js"
 export { Array, Boolean, Date, NonEmptyArray, NullOr, Number, ReadonlyMap, ReadonlySet, Struct, Tuple } from "./ext.js"
